Additions to material abstraction
Materials each have set MaterialLayouts which define various attributes for each material. 2 materials can have the same layout but with different values for the shader parameters (such as textures or float values). More testing needs to be done as well as adding support for textures and layout-wide buffers (if i feels like it).